Fix find_tags_to_copy in images without latest tag
skopeo inspect is used to describe lower level details of a given image.
As a result, it default to the latest tag if none is passed. In images
that don't have a latest tag defined, skopeo will fail.

skopeo list-tags doesn't actually touch on image tags and doesn't suffer
the same problem.